Fix bug in Int.with()

+++ utils.kt
@@ -34,7 +34,7 @@ private fun Int.hasBit(index: Int) = bit(index) != 0
 private fun Int.setBit(index: Int) = or(1 shl index)
 private fun Int.countSignificantBits() = Int.SIZE_BITS - Integer.numberOfLeadingZeros(this)
 private fun Int.oneIndices() = (0 until countSignificantBits()).filter { bit(it) != 0 }
-private infix fun Int.with(that: Int) = (toLong() shl 32) or that.toLong()
+private infix fun Int.with(that: Int) = (toLong() shl 32) or that.toUInt().toLong()
 private infix fun Int.withShort(that: Int) = shl(16) or that
 private fun decode(code: Long) = (code ushr 32).toInt() to code.toInt()
 private fun decodeShort(code: Int) = (code ushr 16) to (code and 0xFFFF)
